weekly alap updates
git-svn-id: file:///usr/local/svn/cpplib/trunk/cftal@781 18305beb-b1e2-0310-bcd8-b14b5419b1d7

+++ x86vec_ins_ext.h
@@ -47,7 +47,7 @@ namespace x86vec {
         template <unsigned _IDX>
         float extract_f32(__m256 r);
         template <unsigned _IDX>
-        __m256 insert_f32(__m256 r, double v);
+        __m256 insert_f32(__m256 r, float v);
 
         template <unsigned _IDX>
         double extract_f64(__m256d r);
