+++ x86vec_v2s64_inl.h
@@ -252,7 +252,8 @@ inline
 x86vec::v2s64
 x86vec::operator~(const v2s64& a)
 {
-        const __m128i all_set = const_v4u32< -1, -1, -1, -1 >::iv();
+        const __m128i all_set = const_v4u32<uint32_t(-1), uint32_t(-1), 
+					    uint32_t(-1), uint32_t(-1)>::iv();
         return _mm_xor_si128(a(), all_set);
 }
